backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Analytics & User Tracking
amplitude-analytics>=1.0.0
mixpanel-python>=4.10.0

# Performance & Caching
redis>=5.0.0
celery>=5.3.0
hiredis>=2.2.0

# Database & ORM Enhancements
sqlalchemy-utils>=0.41.0
sqlalchemy-json>=0.5.0

# API Documentation & Schema
openapi-spec-validator>=0.6.0
fastapi-cors>=0.0.6
orjson>=3.9.0

# Security & Rate Limiting
slowapi>=0.1.8
cryptography>=41.0.0

# Monitoring & Logging
prometheus-client>=0.17.0
python-json-logger>=2.0.0
sentry-sdk>=1.33.0

# Web Scraping & Content
httpx>=0.25.0
beautifulsoup4>=4.12.0

# Task Queue & Async
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx[http2]>=0.25.0
python-multipart>=0.0.6
